+++ src/base/datatoimage.h
@@ -3,4 +3,4 @@
 #include "header.h"
 #include "glm.hpp"
 
-void DataToImage(const glm::vec3* data);
+void DataToImage(const glm::dvec3* data);
